CARDS-1623: Add basic support for DICOM image storage Moved dicomDataDictionary to a separate directory
