+++ modules/data-entry/src/main/frontend/src/dicom/dicomDataDictionary.js
